requirements.txt
flask
flask-cors
pandas
numpy
scipy
scikit-learn
fpdf2
openpyxl